관심사의 분리 : RestaurantInputContainer를 구현한다
